requirements.txt
pyepics==3.5.1
siriuspy==2.46.0 
